+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 2.0.22
 
 - PHP 8.1 compatibility fixes
 - Upgraded to Predis v1.1.9